small refactor: moved moved get_nearest_road_block from amenities.rs to mod.rs
