+++ src/element_processing/mod.rs
@@ -19,6 +19,7 @@ pub mod tree;
 pub mod water_areas;
 pub mod waterways;
 
+use crate::floodfill_cache::RoadMaskBitmap;
 use crate::osm_parser::ProcessedNode;
 
 /// Merges way segments that share endpoints into closed rings.
@@ -119,4 +120,33 @@ pub fn merge_way_segments(rings: &mut Vec<Vec<ProcessedNode>>) {
     if merged_len > 0 {
         merge_way_segments(rings);
     }
+}
+
+/// Looks outward from (x, z) in each of the four cardinal directions,
+/// up to max_radius blocks away, and returns the (x, z) position of
+/// the nearest road node found.
+///
+
+/// Returns None if no road node exists within range.
+/// Callers can use the returned position to derive a facing direction,
+/// compute a distance, or do anything else they need.
+pub fn get_nearest_road_block(
+    x: i32,
+    z: i32,
+    max_radius: i32,
+    road_mask: &RoadMaskBitmap,
+) -> Option<(i32, i32)> {
+    // Begins at 2 and skips to 4, 6, 8, etc.
+    for dist in (2..=max_radius).step_by(2) {
+        // Cross pattern: North, South, West, East
+        let candidates = [(x, z - dist), (x, z + dist), (x - dist, z), (x + dist, z)];
+
+        for (cx, cz) in candidates {
+            if road_mask.contains(cx, cz) {
+                return Some((cx, cz));
+            }
+        }
+    }
+
+    None
 }
